Match gem.cpp
Update Clamp inline, use GX enums and defines, fix pointer type uses, update misc. inlines

+++ include/common/Blitter.h
@@ -56,7 +56,7 @@ struct Blitter_Particle {
     float unk20;
     uint unk24;
     float angle;
-    int unk2C;
+    float unk2C;
 	
 	void Draw(int);
 	void DrawFrames(int, float);
